% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_main.R
\name{checkCurrent}
\alias{checkCurrent}
\title{checkCurrent}
\usage{
checkCurrent(dtCols = c(D157 = "V14"))
}
\arguments{
\item{dtCols}{date columns for each DB}
}
\value{
data.frame
}
\description{
list up all rds files in "./db" folder recursively, and check the last update date.
}
