% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_main.R
\name{cleanUpNonBJD}
\alias{cleanUpNonBJD}
\title{cleanUpNonBJD}
\usage{
cleanUpNonBJD(bjdFn = "./_BJD/BJD.txt")
}
\arguments{
\item{bjdFn}{BJD file name}
}
\description{
Find all rds files in ./db and read them into data frame.
source name is that found file name without extension ".rds".
Loop DB2folder function for all rds files.
Rename the rds files to rds.done.
}
