% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_main.R
\name{DB2folder}
\alias{DB2folder}
\title{DB2folder}
\usage{
DB2folder(db, BJDcol, dbSrc, jpth = "R:/_library/_jijuk/")
}
\arguments{
\item{db}{data.frame}

\item{BJDcol}{BJD column name}

\item{dbSrc}{db source}

\item{jpth}{jijuk path}
}
\description{
Split Jijuk DB by BJD, and save them in the folder made of BJD.
}
