% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_main.R
\name{downloadInfo}
\alias{downloadInfo}
\title{downloadInfo}
\usage{
downloadInfo(download = "./_download(zip)")
}
\arguments{
\item{download}{download folder}
}
\value{
data.frame
}
\description{
This function is to list files in download folder.
}
