% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_main.R
\name{Dwn2standBy2}
\alias{Dwn2standBy2}
\title{Dwn2standBy2}
\usage{
Dwn2standBy2(
  DB,
  pub_time,
  download = "./_download(zip)",
  stand_by = "./_stand_by"
)
}
\arguments{
\item{DB}{DB type}

\item{pub_time}{publication time}

\item{download}{download folder}

\item{stand_by}{stand_by folder}
}
\description{
This function is to filter the files in the download folder by the time and DB type, and unzip the files to the stand_by folder.
If target files are already exist in the stand_by folder, skip them.
If pub_time is specified, the function will download the files that file name contains the pub_time, instead of the time range.
Log the process of the function to .log file.
}
