% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_main.R
\name{lastDate}
\alias{lastDate}
\title{lastDate}
\usage{
lastDate(.inf = info)
}
\arguments{
\item{.inf}{info object}
}
\value{
lastDt
}
\description{
This function returns the last date of the record in the info database.
}
