% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_main.R
\name{stnBy2DB}
\alias{stnBy2DB}
\title{stnBy2DB}
\usage{
stnBy2DB(stand_by = "./_stand_by")
}
\arguments{
\item{stand_by}{stand_by folder}
}
\description{
This function is to list the files in the stand_by folder, and is to split the files into DBs by BJD.
And then splited DBs are saved in the DB folder.
}
